feat: updated role based permissions setup

+++ libs/account-service/src/commands/handlers/RemoveBusinessMemberHandler.ts
@@ -1,12 +1,12 @@
-import { Inject, NotFoundException } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { RemoveBusinessMemberCommand } from '../impl';
-import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
-import { AppLogger } from 'libs/common/src/logger/logger.service';
-import modelsFormatter from 'libs/common/src/middlewares/models.formatter';
-import { Business, BusinessInfo } from 'libs/common/src/models/business.model';
+import { AccountRole } from '@app/common/src/constants/enums';
 import { Account } from '@app/common/src/models/account.model';
+import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
+import { Business } from 'libs/common/src/models/business.model';
+import { AppLogger } from 'libs/common/src/logger/logger.service';
+import { ForbiddenException, Inject, NotFoundException } from '@nestjs/common';
 
 @CommandHandler(RemoveBusinessMemberCommand)
 export class RemoveBusinessMemberHandler implements ICommandHandler<
@@ -19,13 +19,25 @@ export class RemoveBusinessMemberHandler implements ICommandHandler<
     private readonly accountRepository: Repository<Account>,
     @InjectRepository(Business)
     private readonly businessRepository: Repository<Business>,
-  ) {}
+  ) { }
 
   async execute(command: RemoveBusinessMemberCommand) {
     try {
       this.logger.log(`[REMOVE-BUSINESS-MEMBER-HANDLER-PROCESSING]`);
 
       const { accountId, secureUser } = command;
+
+      if (secureUser.role === AccountRole.MEMBER) {
+        throw new ForbiddenException(
+          'You do not have permission to remove members.',
+        );
+      }
+
+      if (secureUser.id === accountId) {
+        throw new ForbiddenException(
+          'You do not have permission to remove yourself.',
+        );
+      }
 
       const business = await this.businessRepository.findOne({
         where: [
@@ -53,6 +65,12 @@ export class RemoveBusinessMemberHandler implements ICommandHandler<
 
       if (!memberToRemove) {
         throw new NotFoundException('Member not found in this business.');
+      }
+
+      if (secureUser.role === AccountRole.ADMIN && memberToRemove.id === business.account.id) {
+        throw new ForbiddenException(
+          'You do not have permission to remove this member.',
+        );
       }
 
       await this.accountRepository.remove(memberToRemove);
